+++ .github/CODEOWNERS
@@ -239,6 +239,7 @@ packages/react-components/react-jsx-runtime @microsoft/teams-prg
 packages/react-components/react-toast @microsoft/teams-prg
 packages/react-components/react-search-preview @microsoft/cxe-coastal
 packages/react-components/react-colorpicker-compat @microsoft/cxe-red @sopranopillow
+packages/react-components/react-anatomy-inspector @microsoft/cxe-red
 # <%= NX-CODEOWNER-PLACEHOLDER %>
 
 ## Components
