+++ packages/react-components/react-anatomy-inspector-preview/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["plugin:@fluentui/eslint-plugin/react"],
+  "root": true
+}
