+++ packages/react-components/react-anatomy-inspector-preview/.npmignore
@@ -1,0 +1,38 @@
+.storybook/
+.vscode/
+bundle-size/
+config/
+coverage/
+docs/
+etc/
+node_modules/
+src/
+stories/
+dist/types/
+temp/
+__fixtures__
+__mocks__
+__tests__
+
+*.api.json
+*.log
+*.spec.*
+*.cy.*
+*.test.*
+*.yml
+
+# config files
+*config.*
+*rc.*
+.editorconfig
+.eslint*
+.git*
+.prettierignore
+.swcrc
+project.json
+
+# exclude gitignore patterns explicitly
+!lib
+!lib-commonjs
+!lib-amd
+!dist/*.d.ts
