+++ packages/react-components/react-anatomy-inspector-preview/README.md
@@ -1,0 +1,5 @@
+# @fluentui/react-anatomy-inspector
+
+**Anatomy Inspector components for [Fluent UI React](https://react.fluentui.dev/)**
+
+These are not production-ready components and **should never be used in product**. This space is useful for testing new components whose APIs might change before final release.
