+++ packages/react-components/react-anatomy-inspector-preview/config/tests.js
@@ -1,0 +1,1 @@
+/** Jest test setup file. */
