+++ packages/react-components/react-anatomy-inspector-preview/just.config.ts
@@ -1,0 +1,5 @@
+import { preset, task } from '@fluentui/scripts-tasks';
+
+preset();
+
+task('build', 'build:react-components').cached?.();
