+++ packages/react-components/react-anatomy-inspector-preview/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "@fluentui/react-anatomy-inspector",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "packages/react-components/react-anatomy-inspector/src",
+  "tags": ["platform:web", "vNext"],
+  "implicitDependencies": []
+}
