+++ packages/react-components/react-anatomy-inspector-preview/src/index.tsx
@@ -1,0 +1,1 @@
+export { AnatomyInspector } from './components/AnatomyInspector';
